[Chore](build) add some compile check (#53637)
add some compile check

+++ be/src/pipeline/exec/join/process_hash_table_probe_impl.h
@@ -161,7 +161,7 @@ void ProcessHashTableProbe<JoinOpType>::probe_side_output_column(vectorized::Mut
 template <int JoinOpType>
 template <typename HashTableType>
 typename HashTableType::State ProcessHashTableProbe<JoinOpType>::_init_probe_side(
-        HashTableType& hash_table_ctx, size_t probe_rows, const uint8_t* null_map) {
+        HashTableType& hash_table_ctx, uint32_t probe_rows, const uint8_t* null_map) {
     // may over batch size 1 for some outer join case
     _probe_indexs.resize(_batch_size + 1);
     _build_indexs.resize(_batch_size + 1);
